style: 유효성 검증기 코드 정리
