+++ packages/validate/fn.ts
@@ -7,7 +7,7 @@ import { isNullable } from '../util/typeCheck';
  * @param val
  * @returns 비어있지 않다면 true, 아니면 false.
  */
-function required(val: string) {
+export function required(val: string) {
   return !isNullable(val) && !!val;
 }
 
@@ -17,8 +17,10 @@ function required(val: string) {
  * @param min 최소 길이
  * @param max 최대 길이
  */
-const length = (min: number, max: number) => (val: string) =>
-  required(val) && val.length >= min && val.length <= max;
+export function length(min: number, max: number) {
+  return (val: string) =>
+    required(val) && val.length >= min && val.length <= max;
+}
 
 /**
  * 이메일 유효성을 검증한다.
@@ -26,62 +28,50 @@ const length = (min: number, max: number) => (val: string) =>
  * 내부적으로 validLength 를 수행하여 7~100 자 까지 인지를 함께 검증한다.
  * @param val
  */
-function email(val: string) {
+export function email(val: string) {
   return (
     length(7, 100)(val) && /^(\w|-|\.)+@\w+([.-]?\w+)*(\.\w{2,5})+$/.test(val)
   );
 }
 
 /**
- * 비밀번호의 유효성을 체크한다. - 약한 버전 - -
+ * 비밀번호의 유효성을 체크한다.
  * 숫자, 영문 각각 1자 이상 포함되고
- * 총 길이가 6자~20자 까지 인지를 확인한다.
+ * 총 길이가 8자~20자 까지 인지를 확인한다.
  * @param val
  */
-function passwordWeak(val: string) {
+export function password(val: string) {
   //  숫자, 영문 대/소문자 및 특수문자가 모두 최소 1자 이상 포함되고
   // return (/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*])(?=.{6,20})/).test(val);
-  return /^(?=.*[A-z])(?=.*[0-9])(?=.{6,20})/.test(val);
-}
-
-/**
- * 비밀번호의 유효성을 체크한다. - 강력 버전 - -
- * 숫자, 영문 대/소문자 및 특수문자가 모두 최소 1자 이상 포함되고
- * 총 길이가 6자~32자 까지 인지를 확인한다.
- * @param val
- */
-function password(val: string) {
-  return /^(?=.*[A-z])(?=.*[~!@#$%^&*()\-=+_';<>/.`:"\\,[\]?|{}])(?=.*[0-9]).{6,32}$/.test(
-    val
-  );
+  return /^(?=.*[A-z])(?=.*[0-9])(?=.{8,20})/.test(val);
 }
 
 /**
  * 소문자가 존재하는지 체크한다.
  * @param val
  */
-function lowercase(val: string) {
+export function lowercase(val: string) {
   return /^(?=.*[a-z])/.test(val);
 }
 /**
  * 대문자가 존재하는지 체크한다.
  * @param val
  */
-function uppercase(val: string) {
+export function uppercase(val: string) {
   return /^(?=.*[A-Z])/.test(val);
 }
 /**
  * 숫자가 존재하는지 체크한다.
  * @param val
  */
-function numbers(val: string) {
+export function numbers(val: string) {
   return /^(?=.*\d)/.test(val);
 }
 /**
  * 특수문자가 존재하는지 체크한다.
  * @param val
  */
-function symbols(val: string) {
+export function symbols(val: string) {
   return /^(?=.*[~!@#$%^&*()\-=+_';<>/.`:"\\,[\]?|{}])/.test(val);
 }
 
@@ -89,7 +79,7 @@ function symbols(val: string) {
  * 한글 또는 영문으로 2글자 이상, 100글자 이하인지 확인한다.
  * @param val 확인할 값
  */
-function korAndEng(val: string) {
+export function korAndEng(val: string) {
   return /^([A-Za-z]|[가-힣]){2,100}$/.test(val);
 }
 
@@ -98,14 +88,14 @@ function korAndEng(val: string) {
  * 값은 번호만 이뤄져 있어야 하며 앞이 0으로 채워져 있어도 유효하다.
  * @param val 확인할 값
  */
-function numberOnly(val: string) {
+export function numberOnly(val: string) {
   return /^[0-9]{1,}$/.test(val);
 }
 
 /**
  * 특정범위의 값인지 여부를 확인한다.
  * ```ts
- * const checker = validate.fn.range(10, 100);
+ * const checker = validateFn.range(10, 100);
  *
  * checker(50); // true
  * checker(9); // false
@@ -115,7 +105,7 @@ function numberOnly(val: string) {
  * @param max
  * @returns
  */
-function range(min: number, max: number) {
+export function range(min: number, max: number) {
   return function innerRange(val: string | number) {
     const num = Number(val);
 
@@ -129,7 +119,7 @@ function range(min: number, max: number) {
  * 중간에 dash (-) 가 있어야 한다.
  * @param val 확인할 값
  */
-function phoneWithDash(val: string) {
+export function phoneWithDash(val: string) {
   return /^(\d{4}-\d{4}|\d{2,3}-\d{3,4}-\d{3,4})$/.test(val);
 }
 
@@ -139,7 +129,7 @@ function phoneWithDash(val: string) {
  *
  * @param val 확인할 값
  */
-function phone(val: string) {
+export function phone(val: string) {
   return length(8, 11)(val) && numbers(val);
 }
 
@@ -147,24 +137,6 @@ function phone(val: string) {
  * 사업자 등록번호 유효성을 검증한다.
  * @param val
  */
-function companyRegNumber(val: string) {
+export function companyRegNumber(val: string) {
   return /^\d{3}-\d{2}-\d{5}$/.test(val);
 }
-
-export default {
-  required,
-  email,
-  passwordWeak,
-  password,
-  lowercase,
-  uppercase,
-  numbers,
-  symbols,
-  korAndEng,
-  numberOnly,
-  range,
-  phoneWithDash,
-  phone,
-  companyRegNumber,
-  length,
-};
